module_5/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "faker>=37.4.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydeps>=3.0.1",
    "pylint>=3.3.7",
    "reportlab>=4.4.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
